+++ src/App.css
@@ -1,3 +1,5 @@
+@import "tailwindcss";
+
 .logo.vite:hover {
   filter: drop-shadow(0 0 2em #747bff);
 }
